+++ CHANGELOG.md
@@ -20,6 +20,7 @@ New `standalone` preset (included in `recommended`) with the following rules:
 - [no-browseranimationsmodule](./docs/standalone/NO_BROWSERANIMATIONSMODULE.md)
 - [no-noopanimationsmodule](./docs/standalone/NO_NOOPANIMATIONSMODULE.md)
 - [no-serviceworkermodule](./docs/standalone/NO_SERVICEWORKERMODULE.md)
+- [no-servermodule](./docs/standalone/NO_SERVERMODULE.md)
 - [no-withmodule-testing](./docs/standalone/NO_WITHMODULE_TESTING.md)
 
 Related rule not in the preset: